+++ internal/kgateway/reports/status.go
@@ -7,6 +7,7 @@ import (
 	"slices"
 	"strings"
 
+	pluginsdkreporter "github.com/kgateway-dev/kgateway/v2/pkg/pluginsdk/reporter"
 	"github.com/solo-io/go-utils/contextutils"
 	"istio.io/istio/pkg/ptr"
 	"k8s.io/apimachinery/pkg/api/meta"
@@ -208,14 +209,14 @@ func parentString(ref gwv1.ParentReference) string {
 // to a given report, i.e. set healthy conditions
 func addMissingGatewayConditions(gwReport *GatewayReport) {
 	if cond := meta.FindStatusCondition(gwReport.GetConditions(), string(gwv1.GatewayConditionAccepted)); cond == nil {
-		gwReport.SetCondition(GatewayCondition{
+		gwReport.SetCondition(pluginsdkreporter.GatewayCondition{
 			Type:   gwv1.GatewayConditionAccepted,
 			Status: metav1.ConditionTrue,
 			Reason: gwv1.GatewayReasonAccepted,
 		})
 	}
 	if cond := meta.FindStatusCondition(gwReport.GetConditions(), string(gwv1.GatewayConditionProgrammed)); cond == nil {
-		gwReport.SetCondition(GatewayCondition{
+		gwReport.SetCondition(pluginsdkreporter.GatewayCondition{
 			Type:   gwv1.GatewayConditionProgrammed,
 			Status: metav1.ConditionTrue,
 			Reason: gwv1.GatewayReasonProgrammed,
@@ -229,28 +230,28 @@ func addMissingGatewayConditions(gwReport *GatewayReport) {
 func addMissingListenerConditions(lisReport *ListenerReport) {
 	// set healthy conditions for Condition Types not set yet (i.e. no negative status yet, we can assume positive)
 	if cond := meta.FindStatusCondition(lisReport.Status.Conditions, string(gwv1.ListenerConditionAccepted)); cond == nil {
-		lisReport.SetCondition(ListenerCondition{
+		lisReport.SetCondition(pluginsdkreporter.ListenerCondition{
 			Type:   gwv1.ListenerConditionAccepted,
 			Status: metav1.ConditionTrue,
 			Reason: gwv1.ListenerReasonAccepted,
 		})
 	}
 	if cond := meta.FindStatusCondition(lisReport.Status.Conditions, string(gwv1.ListenerConditionConflicted)); cond == nil {
-		lisReport.SetCondition(ListenerCondition{
+		lisReport.SetCondition(pluginsdkreporter.ListenerCondition{
 			Type:   gwv1.ListenerConditionConflicted,
 			Status: metav1.ConditionFalse,
 			Reason: gwv1.ListenerReasonNoConflicts,
 		})
 	}
 	if cond := meta.FindStatusCondition(lisReport.Status.Conditions, string(gwv1.ListenerConditionResolvedRefs)); cond == nil {
-		lisReport.SetCondition(ListenerCondition{
+		lisReport.SetCondition(pluginsdkreporter.ListenerCondition{
 			Type:   gwv1.ListenerConditionResolvedRefs,
 			Status: metav1.ConditionTrue,
 			Reason: gwv1.ListenerReasonResolvedRefs,
 		})
 	}
 	if cond := meta.FindStatusCondition(lisReport.Status.Conditions, string(gwv1.ListenerConditionProgrammed)); cond == nil {
-		lisReport.SetCondition(ListenerCondition{
+		lisReport.SetCondition(pluginsdkreporter.ListenerCondition{
 			Type:   gwv1.ListenerConditionProgrammed,
 			Status: metav1.ConditionTrue,
 			Reason: gwv1.ListenerReasonProgrammed,
@@ -263,14 +264,14 @@ func addMissingListenerConditions(lisReport *ListenerReport) {
 // to a given report, i.e. set healthy conditions
 func addMissingParentRefConditions(report *ParentRefReport) {
 	if cond := meta.FindStatusCondition(report.Conditions, string(gwv1.RouteConditionAccepted)); cond == nil {
-		report.SetCondition(RouteCondition{
+		report.SetCondition(pluginsdkreporter.RouteCondition{
 			Type:   gwv1.RouteConditionAccepted,
 			Status: metav1.ConditionTrue,
 			Reason: gwv1.RouteReasonAccepted,
 		})
 	}
 	if cond := meta.FindStatusCondition(report.Conditions, string(gwv1.RouteConditionResolvedRefs)); cond == nil {
-		report.SetCondition(RouteCondition{
+		report.SetCondition(pluginsdkreporter.RouteCondition{
 			Type:   gwv1.RouteConditionResolvedRefs,
 			Status: metav1.ConditionTrue,
 			Reason: gwv1.RouteReasonResolvedRefs,
